Refactor DXGI impl. into separate cpps.
In the future, DXGI will be conditionally built for projects that don't require DXGI.

Issue: #683
