+++ src/gpgmm/BUILD.gn
@@ -115,6 +115,8 @@ source_set("gpgmm_sources") {
     sources += [
       "d3d12/BudgetUpdateD3D12.cpp",
       "d3d12/BudgetUpdateD3D12.h",
+      "d3d12/BudgetUpdateDXGI.cpp",
+      "d3d12/BudgetUpdateDXGI.h",
       "d3d12/BufferAllocatorD3D12.cpp",
       "d3d12/BufferAllocatorD3D12.h",
       "d3d12/CapsD3D12.cpp",
@@ -136,6 +138,8 @@ source_set("gpgmm_sources") {
       "d3d12/ResidencyListD3D12.h",
       "d3d12/ResidencyManagerD3D12.cpp",
       "d3d12/ResidencyManagerD3D12.h",
+      "d3d12/ResidencyManagerDXGI.cpp",
+      "d3d12/ResidencyManagerDXGI.h",
       "d3d12/ResourceAllocationD3D12.cpp",
       "d3d12/ResourceAllocationD3D12.h",
       "d3d12/ResourceAllocationTrackingAllocatorD3D12.cpp",
