+++ src/gpgmm/d3d12/ResidencyManagerD3D12.cpp
@@ -27,6 +27,7 @@
 #include "gpgmm/d3d12/LogD3D12.h"
 #include "gpgmm/d3d12/ResidencyHeapD3D12.h"
 #include "gpgmm/d3d12/ResidencyListD3D12.h"
+#include "gpgmm/d3d12/ResidencyManagerDXGI.h"
 #include "gpgmm/d3d12/UtilsD3D12.h"
 #include "gpgmm/utils/Math.h"
 
@@ -921,39 +922,6 @@ namespace gpgmm::d3d12 {
 
     ID3D12Device* ResidencyManager::GetDevice() const {
         return mDevice;
-    }
-
-    // ResidencyManagerDXGI
-
-    ResidencyManagerDXGI::ResidencyManagerDXGI(const RESIDENCY_MANAGER_DESC& descriptor,
-                                               ID3D12Device* pDevice,
-                                               IDXGIAdapter3* pAdapter,
-                                               std::unique_ptr<Caps> caps)
-        : ResidencyManager(descriptor, pDevice, std::move(caps)), mAdapter(pAdapter) {
-        ASSERT(mAdapter != nullptr);
-    }
-
-    ResidencyManagerDXGI::~ResidencyManagerDXGI() = default;
-
-    HRESULT ResidencyManagerDXGI::QueryMemoryInfoImpl(const RESIDENCY_HEAP_SEGMENT& heapSegment,
-                                                      RESIDENCY_MEMORY_INFO* pMemoryInfoOut) {
-        // Residency heap segments are 1:1 with DXGI memory segment groups.
-        DXGI_QUERY_VIDEO_MEMORY_INFO queryVideoMemoryInfoOut;
-        GPGMM_RETURN_IF_FAILED(mAdapter->QueryVideoMemoryInfo(
-            0, static_cast<DXGI_MEMORY_SEGMENT_GROUP>(heapSegment), &queryVideoMemoryInfoOut));
-        pMemoryInfoOut->AvailableForReservation = queryVideoMemoryInfoOut.AvailableForReservation;
-        pMemoryInfoOut->Budget = queryVideoMemoryInfoOut.Budget;
-        pMemoryInfoOut->CurrentReservation = queryVideoMemoryInfoOut.CurrentReservation;
-        pMemoryInfoOut->CurrentUsage = queryVideoMemoryInfoOut.CurrentUsage;
-        return S_OK;
-    }
-
-    std::shared_ptr<BudgetUpdateTask> ResidencyManagerDXGI::CreateBudgetUpdateTask() {
-        return std::make_shared<BudgetUpdateTaskDXGI>(this);
-    }
-
-    IDXGIAdapter3* ResidencyManagerDXGI::GetAdapter() const {
-        return mAdapter;
     }
 
 }  // namespace gpgmm::d3d12
